refactor: 관리자 결제 통계 응답 DTO 분리 및 배치 정리

+++ src/main/java/com/byunsum/ticket_reservation/payment/dto/PaymentStatistics.java
@@ -1,7 +1,9 @@
 package com.byunsum.ticket_reservation.payment.dto;
 
+import java.math.BigDecimal;
+
 public interface PaymentStatistics {
     String getPaymentMethod();
     Long getCount();
-    Long getTotal();
+    BigDecimal getTotal();
 }
